% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{getNumVectorsHutchinson}
\alias{getNumVectorsHutchinson}
\title{Compute the number of vectors used for trace estimation}
\usage{
getNumVectorsHutchinson(e, d)
}
\arguments{
\item{e}{`numeric` denoting the relative error margin}

\item{d}{`numeric`. 1-d is the probability the the relative error is bounded
by e.}
}
\value{
a scalar
}
\description{
Follows Skorski, M. (2021). Modern Analysis of Hutchinson’s Trace Estimator.
2021 55th Annual Conference on Information Sciences and Systems (CISS), 1–5.
https://doi.org/10.1109/CISS50987.2021.9400306
}
