% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSpam.R
\name{importLinv}
\alias{importLinv}
\title{Generate spam object from L inverse file}
\usage{
importLinv(pth)
}
\arguments{
\item{pth}{path to matrix market file for L inverse matrix in dgTMatrix format}
}
\value{
A \code{spam} sparse matrix
}
\description{
RandPedPCA relies in the \code{spam} onject format. But matrices are commonly
stored in other formats.
}
