% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedLinv}
\alias{pedLinv}
\title{Example pedigree L inverse matrix}
\format{
## `pedLinv`
Matrix object of class 'spam' of dimension 7100x7100,
    with 21040 (row-wise) nonzero elements.
    Density of the matrix is 0.0417%.
    Class 'spam' (32-bit)
}
\source{
Simulation
}
\usage{
pedLinv
}
\description{
An L inverse matrix generated from an AlphaSimR simulation of 50 generations.
An original population splits into sub-populations A and B. After a number of
generations, crossbreeding starts.
}
\keyword{datasets}
