% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedLInv2}
\alias{pedLInv2}
\title{L inverse matrix of the two-population example pedigree}
\format{
## `pedLInv2`
Matrix object of class 'spam' of dimension 2650x2650,
    with 7750 (row-wise) nonzero elements.
    Density of the matrix is 0.11%.
    Class 'spam' (32-bit)
}
\source{
Simulation
}
\usage{
pedLInv2
}
\description{
An L inverse matrix generated from an AlphaSimR simulation of 20 generations.
Two diverged populations A and B. After a number of
generations, crossbreeding starts.
}
\keyword{datasets}
