% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedLInv4}
\alias{pedLInv4}
\title{L inverse matrix of the four-population example pedigree}
\format{
## `pedLInv4`
Matrix object of class 'spam' of dimension 4200x4200,
    with 12200 (row-wise) nonzero elements.
    Density of the matrix is 0.0692%.
    Class 'spam' (32-bit)
}
\source{
Simulation
}
\usage{
pedLInv4
}
\description{
An L inverse matrix generated from an AlphaSimR simulation of 20 generations.
One population, ABCD, is split into four, A, B, C, D.
}
\keyword{datasets}
