% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedMeta2}
\alias{pedMeta2}
\title{Metadata associated with the 2nd example pedigree}
\format{
## `pedMeta2`
A 'data.frame' of 2650 individuals (rows) with 12 variables (cols):
\describe{
  \item{id}{Integer individual ID}
  \item{population}{Population code. A, B or AB}
  \item{generation}{Generation of the individual}
  \item{mid}{dam ID}
  \item{fid}{sire ID}
  \item{gv1}{genetic value}
  \item{pv1}{phenotypic value}
  \item{gv2}{genetic value}
  \item{pv2}{phenotypic value}
  \item{gv}{genetic value}
  \item{pv}{phenotypic value}
  \item{generationPlotShift}{for plotting}
}
}
\source{
Simulation
}
\usage{
pedMeta2
}
\description{
A dataframe.
}
\keyword{datasets}
