% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{3D plot using rgl}
\usage{
plot3D(x, dims = c(1, 2, 3), ...)
}
\arguments{
\item{x, }{an rppca object}

\item{dims, }{vector of length 3 - indices of the PCs to plot}

\item{...}{additional arguments passed to rgl::plot3d}
}
\description{
A simple wrapper around rgl's pot3d function.
}
\details{
Note, different to `plot.rppca`, which is relatively slow, `plot3D` does
not down-sample the principal components and it ignores the `ds` slot of an
`rppca` object if present.
}
