% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{randTraceHutchinson}
\alias{randTraceHutchinson}
\title{Trace estimation for sparse L inverse matrices}
\usage{
randTraceHutchinson(L, numVectors)
}
\arguments{
\item{L}{A pedigree's L inverse matrix}

\item{`numVectors`, }{an `integer` specifying how many random vectors to use

The higher `numVectors`, the higher the accuracy and the longer the runtime.
Accuracy can be estimated with the function `getNumVectorsHutchinson`.}
}
\value{
a scalar
}
\description{
Using Hutchinson's method
}
