% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSpam.R
\name{sparse2spam}
\alias{sparse2spam}
\title{Convert generic sparse matrix to spam format}
\usage{
sparse2spam(sprs)
}
\arguments{
\item{sprs}{A sparse matrix.}
}
\value{
A \code{spam} sparse matrix
}
\description{
Convert generic sparse matrix to spam format
}
