% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedLInv}
\alias{pedLInv}
\title{L inverse matrix of the one-population example pedigree}
\format{
## `pedLInv`
Matrix object of class 'spam' of dimension 2100x2100,
    with 6100 (row-wise) nonzero elements.
    Density of the matrix is 0.138%.
Class 'spam' (32-bit)
}
\source{
Simulation
}
\usage{
pedLInv
}
\description{
An L inverse matrix generated from an AlphaSimR simulation of 20 generations.
}
\keyword{datasets}
