% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedMeta}
\alias{pedMeta}
\title{Metadata associated with one-population example}
\format{
## `pedMeta`
A 'data.frame' of 2100 individuals (rows) with 9 variables (cols):
\describe{
  \item{id}{Integer individual ID}
  \item{population}{Population code. A, B or AB}
  \item{generation}{Generation of the individual}
  \item{mid}{dam ID}
  \item{fid}{sire ID}
  \item{gv1}{genetic value}
  \item{pv1}{phenotypic value}
  \item{gv2}{genetic value}
  \item{pv2}{phenotypic value}
}
}
\source{
Simulation
}
\usage{
pedMeta
}
\description{
A dataframe.
}
\keyword{datasets}
