% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{randRangeFinder}
\alias{randRangeFinder}
\title{Generate range matrix}
\usage{
randRangeFinder(L, rank, depth, numVectors)
}
\description{
The efficiency of RandPedPCA relies on the fact that a relatively small
matrix is decomposed instead of the original large L inverse. This function
generates the range matrix.
}
