% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{randRangeFinder}
\alias{randRangeFinder}
\title{Generat range matrix for SVD}
\usage{
randRangeFinder(L, rank, depth, numVectors)
}
\arguments{
\item{L}{a pedigree's L inverse matrix in sparse 'spam' format}

\item{rank}{`integer` how many principal components to return}

\item{depth}{`integer` number of iterations for generating the range matrix}

\item{numVectors}{`integer` > `rank` to specify the oversampling for the}
}
\value{
The range matrix for `randSVD`
}
\description{
Generat range matrix for SVD
}
