% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{randSVD}
\alias{randSVD}
\title{Singular value decomposition in sparse triangular matrix}
\usage{
randSVD(L, rank, depth, numVectors, cent = F)
}
\arguments{
\item{L}{a pedigree's L inverse matrix in sparse 'spam' format}

\item{rank}{`integer` how many principal components to return}

\item{depth}{`integer` number of iterations for generating the range matrix}

\item{numVectors}{`integer` > `rank` to specify the oversampling for the}

\item{cent}{`logical` whether or not to (implicitly) centre the additive
relationship matrix}
}
\value{
A list of three: u (=U), d (=Sigma), and v (=W^T)
}
\description{
Uses random linear algebra, see Halko et al. (2010). Singular value
decomposition (SVD) decomposes a matrix \eqn{X=U\Sigma W^T}
}
