% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{randTraceHutchinson}
\alias{randTraceHutchinson}
\title{Trace estimation for sparse L inverse matrices}
\usage{
randTraceHutchinson(L, numVectors)
}
\arguments{
\item{L}{A pedigree's L inverse matrix}

\item{numVectors, }{an \code{integer} specifying how many random vectors to use

If you do not have a good reason to do otherwise, use the function \code{hutchpp} instead.

The higher \code{numVectors}, the higher the accuracy and the longer the running time.
Accuracy can be estimated with the function \code{getNumVectorsHutchinson}.}
}
\value{
a scalar
}
\description{
Using Hutchinson's method
}
