% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppca.R
\name{rppca}
\alias{rppca}
\title{PCA of pedigree L inverse sparse matrix}
\usage{
rppca(L, method = "randSVD", rank = 10, depth = 3, numVectors = 15)
}
\arguments{
\item{L}{a pedigree's L inverse matrix in sparse `spam` format}

\item{method}{`string` only randSVD (the default) is implemented}

\item{rank}{`integer` how many principal components to return}

\item{depth}{`integer` number of iterations for generating the range matrix}

\item{numVectors}{`integer` > `rank` to specify the oversampling for the
range matrix}
}
\value{
A `list` of one element, `scores`, the principal components
}
\description{
PCA of pedigree L inverse sparse matrix
}
